+++ docs/changelog.md
@@ -1,4 +1,9 @@
 # TBD
+### Fixes
+* No longer adds 300s to the CL genesis timestamp
+
+### Changes
+* Made Docker image more compatible with Kurtosis
 
 # 0.1.3
 ### Changes